refactor: cleaned up models dir fix: removed dateReceived/Completed from StaffForm, replaced with DateCreated
