+++ lib/models/forms/student_form.dart
@@ -1,7 +1,8 @@
 import 'dart:io';
-import 'interfaces/IGenericForm.dart';
+import 'package:athlete_surveyor/models/interfaces/IGenericForm.dart';
 import 'package:athlete_surveyor/models/question.dart';
 import 'package:uuid/uuid.dart';
+
 
 /// A concrete implementation of [IGenericForm], tailored for what students will see.
 /// This includes the loading of a form to *take* a survey, as well as the ability to 
@@ -13,8 +14,8 @@ import 'package:uuid/uuid.dart';
 /// and cached to the device as a draft FormResult so that the student can return to finish 
 /// questions later. 
 /// 
-/// The nested [ResponseType] is intended be read by a service on the Form Editor page that 
-/// would allow the alteration of the desired format. 
+/// The nested [ResponseType] is intended be read by something on the UI layer to provide
+/// a widget based on the question being viewed.
 class StudentForm implements IGenericForm {
 
   final String formName;        // from super
@@ -24,9 +25,9 @@ class StudentForm implements IGenericForm {
   final DateTime? formDateCompleted;
   
   /// TODO: this should be populated by the form_file_service
-  List<File>? attachments = [];
+  @override
+  List<File>? attachments;
 
-  // Map<Question, List<Response>> used later in analytics, ABOVE this scope
   List<Question> questions = [];
 
   /// Constructs a new StudentForm
@@ -49,4 +50,5 @@ class StudentForm implements IGenericForm {
     // Load form data from a remote source or local cache
     // This would populate the questions list with existing data, including drafts
   }
+  
 }
