+++ lib/pages/previous_forms_page.dart
@@ -1,8 +1,8 @@
 // ignore_for_file: library_private_types_in_public_api
-import 'package:athlete_surveyor/models/individual_form_examination_model.dart';
+import 'package:athlete_surveyor/models/forms/individual_form_examination_model.dart';
 import 'package:athlete_surveyor/pages/individual_form_examination_page.dart';
 import 'package:flutter/material.dart';
-import 'package:athlete_surveyor/models/previous_forms_model.dart';
+import 'package:athlete_surveyor/models/forms/previous_forms_model.dart';
 
 //
 class PreviousFormsPage extends StatefulWidget
